新增FLAC格式音乐标签信息写入与封面嵌入
