+++ publish/changeLog.md
@@ -1,6 +1,6 @@
 ### 新增
 
-- 新增FLAC格式音乐标签信息写入
+- 新增FLAC格式音乐标签信息写入与封面嵌入
 
 ### 优化
