+++ src/main/utils/flac-metadata/lib/MetaDataBlock.js
