+++ src/main/utils/flac-metadata/lib/MetaDataBlockPicture.js
