+++ src/main/utils/flac-metadata/lib/MetaDataBlockStreamInfo.js
