+++ src/main/utils/flac-metadata/lib/MetaDataBlockVorbisComment.js
