+++ src/renderer/lang/cnt/view/setting.json
@@ -54,7 +54,7 @@
   "download_name_title": "下載歌曲時的命名方式",
   "download_name": "文件命名方式",
   "download_embed_pic_title": "是否將封面嵌入音頻文件中",
-  "download_embed_pic": "封面嵌入（只支持MP3格式）",
+  "download_embed_pic": "封面嵌入",
   "download_lyric_title": "是否同時下載歌詞文件",
   "download_lyric": "歌詞下載",
   "download_name1": "歌名 - 歌手",
